+++ frontend/src/queries/templates/useTemplateListQuery.ts
@@ -1,7 +1,7 @@
 import { keepPreviousData, useQuery } from '@tanstack/react-query';
 
 import { PAGE_SIZE, QUERY_KEY, getTemplateList, DEFAULT_SORTING_OPTION } from '@/api';
-import { useAuth } from '@/hooks/authentication/useAuth';
+import { useAuth } from '@/hooks/authentication';
 import type { TemplateListResponse, SortingKey } from '@/types';
 
 interface Props {
